feat: Reimplementation of process
This change completely reworks the logic that processes raw nodes into
processed ones. It removes some unnecessary datastructures used to
encapsulate DNS name supersets, and implements a more robust algorithm
for matching raw node (locators) to linkable nodes, roughly:
+ Match nodes with no link ID (locators) to nodes with a link ID (nodes)
where the set of DNS names for the locator is a subset of that for the node.
+ Match unmatched locators to nodes where the set of DNS names for the
locator is a subset of the *nodes DNS superset*.
+ Warn in any locators are left unmatched.